Changed cache hit/miss counters to `long` for wider range support and added instrumentation for `getMetaClass` calls with thread-safe counting.

+++ src/DatenMeister.Core/Runtime/ExtentUrlNavigator.cs
@@ -28,9 +28,9 @@ public class ExtentUrlNavigator(IUriExtent extent, IScopeStorage? scopeStorage)
 
     private static int _cacheMiss = 0;
     
-    public static int CacheHit => _cacheHit;
+    public static long CacheHit => _cacheHit;
 
-    public static int CacheMiss => _cacheMiss;
+    public static long CacheMiss => _cacheMiss;
 
     private ResolveHookContainer? GetResolveHooks()
     {
